Rename logstash-logback-encoder package and artifact
