+++ logbook-logstash/pom.xml
@@ -6,7 +6,7 @@
         <artifactId>logbook-parent</artifactId>
         <version>2.0.0-SNAPSHOT</version>
     </parent>
-    <artifactId>logbook-logstash-logback-encoder</artifactId>
+    <artifactId>logbook-logstash</artifactId>
     <name>Logbook: logstash-logback-encoder</name>
     <description>logstash-logback-encoder support for request and response logging</description>
     <scm>
