+++ logbook-logstash/src/main/java/org/zalando/logbook/logstash/package-info.java
@@ -1,4 +1,4 @@
 @ParametersAreNonnullByDefault
-package org.zalando.logbook.lle;
+package org.zalando.logbook.logstash;
 
 import javax.annotation.ParametersAreNonnullByDefault;
